+++ src/promptflow/promptflow/_sdk/_constants.py
@@ -68,6 +68,7 @@ DEFAULT_VAR_ID = "default_variant_id"
 FLOW_TOOLS_JSON = "flow.tools.json"
 FLOW_META_JSON = "flow.json"
 FLOW_TOOLS_JSON_GEN_TIMEOUT = 60
+FLOW_META_JSON_GEN_TIMEOUT = 60
 PROMPT_FLOW_RUNS_DIR_NAME = ".runs"
 PROMPT_FLOW_EXP_DIR_NAME = ".exps"
 SERVICE_CONFIG_FILE = "pf.yaml"
@@ -76,6 +77,7 @@ PF_SERVICE_PORT_FILE = "pfs.port"
 PF_SERVICE_LOG_FILE = "pfs.log"
 PF_SERVICE_HOUR_TIMEOUT = 1
 PF_SERVICE_MONITOR_SECOND = 60
+PF_SERVICE_WORKER_NUM = 16
 PF_TRACE_CONTEXT = "PF_TRACE_CONTEXT"
 PF_SERVICE_DEBUG = "PF_SERVICE_DEBUG"
 
@@ -254,6 +256,8 @@ class FlowRunProperties:
     # Experiment command node fields only
     COMMAND = "command"
     OUTPUTS = "outputs"
+    RESUME_FROM = "resume_from"
+    COLUMN_MAPPING = "column_mapping"
 
 
 class CommonYamlFields: