+++ src/promptflow/promptflow/_sdk/_errors.py
@@ -81,6 +81,19 @@ class ConnectionNotFoundError(SDKError):
     pass
 
 
+class RequiredEnvironmentVariablesNotSetError(SDKError):
+    """Exception raised if connection from_env required env vars not found."""
+
+    def __init__(self, env_vars: list, cls_name: str):
+        super().__init__(f"Required environment variables {env_vars} to build {cls_name} not set.")
+
+
+class ConnectionNameNotSetError(SDKError):
+    """Exception raised if connection not set when create or update."""
+
+    pass
+
+
 class InvalidRunError(SDKError):
     """Exception raised if run name is not legal."""
 
@@ -89,6 +102,12 @@ class InvalidRunError(SDKError):
 
 class GenerateFlowToolsJsonError(SDKError):
     """Exception raised if flow tools json generation failed."""
+
+    pass
+
+
+class GenerateFlowMetaJsonError(SDKError):
+    """Exception raised if flow json generation failed."""
 
     pass