+++ src/promptflow/promptflow/_sdk/_service/app.py
@@ -7,11 +7,16 @@ import time
 from datetime import datetime, timedelta
 from logging.handlers import RotatingFileHandler
 
-from flask import Blueprint, Flask, g, jsonify, request
+from flask import Blueprint, Flask, current_app, g, jsonify, request
 from flask_cors import CORS
 from werkzeug.exceptions import HTTPException
 
-from promptflow._sdk._constants import PF_SERVICE_HOUR_TIMEOUT, PF_SERVICE_LOG_FILE, PF_SERVICE_MONITOR_SECOND
+from promptflow._sdk._constants import (
+    HOME_PROMPT_FLOW_DIR,
+    PF_SERVICE_HOUR_TIMEOUT,
+    PF_SERVICE_LOG_FILE,
+    PF_SERVICE_MONITOR_SECOND,
+)
 from promptflow._sdk._service import Api
 from promptflow._sdk._service.apis.collector import trace_collector
 from promptflow._sdk._service.apis.connection import api as connection_api
@@ -26,7 +31,7 @@ from promptflow._sdk._service.utils.utils import (
     get_port_from_config,
     kill_exist_service,
 )
-from promptflow._sdk._utils import get_promptflow_sdk_version, overwrite_null_std_logger
+from promptflow._sdk._utils import get_promptflow_sdk_version, overwrite_null_std_logger, read_write_by_user
 from promptflow._utils.thread_utils import ThreadWithContextVars
 
 overwrite_null_std_logger()
@@ -35,6 +40,9 @@ overwrite_null_std_logger()
 def heartbeat():
     response = {"promptflow": get_promptflow_sdk_version()}
     return jsonify(response)
+
+
+CREATED_BY_FOR_LOCAL_TO_CLOUD_TRACE = {}
 
 
 def create_app():
@@ -66,13 +74,44 @@ def create_app():
         # Enable log
         app.logger.setLevel(logging.INFO)
         # each env will have its own log file
-        log_file = get_current_env_pfs_file(PF_SERVICE_LOG_FILE)
+        if sys.executable.endswith("pfcli.exe"):
+            log_file = HOME_PROMPT_FLOW_DIR / PF_SERVICE_LOG_FILE
+            log_file.touch(mode=read_write_by_user(), exist_ok=True)
+        else:
+            log_file = get_current_env_pfs_file(PF_SERVICE_LOG_FILE)
         # Create a rotating file handler with a max size of 1 MB and keeping up to 1 backup files
         handler = RotatingFileHandler(filename=log_file, maxBytes=1_000_000, backupCount=1)
         formatter = logging.Formatter("[%(asctime)s][%(name)s][%(levelname)s] - %(message)s")
         handler.setFormatter(formatter)
         # Set app logger to the only one RotatingFileHandler to avoid duplicate logs
         app.logger.handlers = [handler]
+
+        def initialize_created_by_info():
+            from promptflow._sdk._configuration import Configuration
+            from promptflow._sdk._utils import extract_workspace_triad_from_trace_provider
+
+            trace_provider = Configuration.get_instance().get_trace_provider()
+            if trace_provider is None or extract_workspace_triad_from_trace_provider(trace_provider) is None:
+                return
+            try:
+                import jwt
+                from azure.identity import DefaultAzureCredential
+
+                from promptflow.azure._utils.general import get_arm_token
+
+                default_credential = DefaultAzureCredential()
+
+                token = get_arm_token(credential=default_credential)
+                decoded_token = jwt.decode(token, options={"verify_signature": False})
+                user_object_id, user_tenant_id = decoded_token["oid"], decoded_token["tid"]
+                CREATED_BY_FOR_LOCAL_TO_CLOUD_TRACE.update(
+                    {
+                        "object_id": user_object_id,
+                        "tenant_id": user_tenant_id,
+                    }
+                )
+            except Exception as e:
+                current_app.logger.error(f"Failed to get created_by info, ignore it: {e}")
 
         # Basic error handler
         @api.errorhandler(Exception)
@@ -118,9 +157,6 @@ def create_app():
         def monitor_request():
             while True:
                 time.sleep(PF_SERVICE_MONITOR_SECOND)
-                # For python scenario, since we start waitress in cli, there will be two app. The one used to log in
-                # the parent process will have no "last_request_time" in app.config since the app doesn't run.
-                app.logger.info(f"Promptflow service is running, version: {get_promptflow_sdk_version()}")
                 if "last_request_time" in app.config and datetime.now() - app.config["last_request_time"] > timedelta(
                     hours=PF_SERVICE_HOUR_TIMEOUT
                 ):
@@ -131,6 +167,7 @@ def create_app():
                         kill_exist_service(port)
                     break
 
+        initialize_created_by_info()
         if not sys.executable.endswith("pfcli.exe"):
             monitor_thread = ThreadWithContextVars(target=monitor_request, daemon=True)
             monitor_thread.start()