+++ src/promptflow/promptflow/executor/_service/utils/process_utils.py
@@ -10,12 +10,16 @@ import os
 from datetime import datetime, timedelta
 from typing import Callable
 
+import psutil
+
 from promptflow._core._errors import UnexpectedError
 from promptflow._core.operation_context import OperationContext
 from promptflow._utils.exception_utils import ExceptionPresenter, JsonSerializedPromptflowException
 from promptflow._utils.logger_utils import service_logger
+from promptflow._utils.process_utils import block_terminate_signal_to_parent
 from promptflow.exceptions import ErrorTarget
-from promptflow.executor._service._errors import ExecutionTimeoutError
+from promptflow.executor._service._errors import ExecutionCanceledError, ExecutionTimeoutError
+from promptflow.executor._service.utils.process_manager import ProcessManager
 
 LONG_WAIT_TIMEOUT = timedelta(days=1).total_seconds()
 SHORT_WAIT_TIMEOUT = 10
@@ -26,6 +30,7 @@ async def invoke_sync_function_in_process(
     *,
     args: tuple = (),
     kwargs: dict = {},
+    run_id: str = None,
     context_dict: dict = None,
     wait_timeout: int = LONG_WAIT_TIMEOUT,
 ):
@@ -39,33 +44,52 @@ async def invoke_sync_function_in_process(
         )
         p.start()
         service_logger.info(f"[{os.getpid()}--{p.pid}] Start process to execute the request.")
+        if run_id:
+            ProcessManager().start_process(run_id, p.pid)
 
-        # Wait for the process to finish or timeout asynchronously
-        start_time = datetime.utcnow()
-        while (datetime.utcnow() - start_time).total_seconds() < wait_timeout and p.is_alive():
-            await asyncio.sleep(1)
+        try:
+            # Wait for the process to finish or timeout asynchronously
+            start_time = datetime.utcnow()
+            while (datetime.utcnow() - start_time).total_seconds() < wait_timeout and _is_process_alive(p):
+                await asyncio.sleep(1)
 
-        # Terminate the process if it is still alive after timeout
-        if p.is_alive():
-            service_logger.error(f"[{p.pid}] Stop process for exceeding {wait_timeout} seconds.")
-            p.terminate()
-            p.join()
-            raise ExecutionTimeoutError(wait_timeout)
+            # If process_id is None, it indicates that the process has been terminated by cancel request.
+            if run_id and not ProcessManager().get_process(run_id):
+                raise ExecutionCanceledError(run_id)
 
-        # Raise exception if the process exit code is not 0
-        if p.exitcode != 0:
-            exception = error_dict.get("error", None)
-            if exception is None:
-                raise UnexpectedError(
-                    message="Unexpected error occurred while executing the request",
-                    target=ErrorTarget.EXECUTOR,
-                )
-            # JsonSerializedPromptflowException will be raised here
-            # no need to change to PromptflowException since it will be handled in app.exception_handler
-            raise exception
+            # Terminate the process if it is still alive after timeout
+            if p.is_alive():
+                service_logger.error(f"[{p.pid}] Stop process for exceeding {wait_timeout} seconds.")
+                p.terminate()
+                p.join()
+                raise ExecutionTimeoutError(wait_timeout)
 
-        service_logger.info(f"[{p.pid}] Process finished.")
-        return return_dict.get("result", {})
+            # Raise exception if the process exit code is not 0
+            if p.exitcode != 0:
+                exception = error_dict.get("error", None)
+                if exception is None:
+                    raise UnexpectedError(
+                        message="Unexpected error occurred while executing the request",
+                        target=ErrorTarget.EXECUTOR,
+                    )
+                # JsonSerializedPromptflowException will be raised here
+                # no need to change to PromptflowException since it will be handled in app.exception_handler
+                raise exception
+
+            service_logger.info(f"[{p.pid}--{os.getpid()}] Process finished.")
+            return return_dict.get("result", {})
+        finally:
+            if run_id:
+                ProcessManager().remove_process(run_id)
+
+
+def _is_process_alive(p: multiprocessing.Process):
+    if psutil.pid_exists(p.pid):
+        if psutil.Process(p.pid).status() != psutil.STATUS_ZOMBIE:
+            return True
+    # Call p.join() to clear the zombie process correctly.
+    p.join()
+    return False
 
 
 def _execute_target_function(
@@ -76,6 +100,7 @@ def _execute_target_function(
     error_dict: dict,
     context_dict: dict,
 ):
+    block_terminate_signal_to_parent()
     with exception_wrapper(error_dict):
         if context_dict:
             OperationContext.get_instance().update(context_dict)