+++ src/promptflow/promptflow/tracing/__init__.py
@@ -1,0 +1,9 @@
+# ---------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# ---------------------------------------------------------
+
+__path__ = __import__("pkgutil").extend_path(__path__, __name__)  # type: ignore
+
+from ._start_trace import start_trace
+
+__all__ = ["start_trace"]