+++ src/promptflow/tests/executor/unittests/executor/_service/apis/test_execution.py
@@ -1,31 +1,10 @@
-import uuid
 from pathlib import Path
-from tempfile import mkdtemp
 from unittest.mock import patch
 
 import pytest
 from fastapi.testclient import TestClient
 
-from .....utils import get_flow_folder, load_content
-
-
-def construct_flow_execution_request_json(flow_folder, inputs=None, connections=None):
-    working_dir = get_flow_folder(flow_folder)
-    tmp_dir = Path(mkdtemp())
-    log_path = tmp_dir / "log.txt"
-    return {
-        "run_id": str(uuid.uuid4()),
-        "working_dir": working_dir.as_posix(),
-        "flow_file": "flow.dag.yaml",
-        "output_dir": tmp_dir.as_posix(),
-        "connections": connections,
-        "log_path": log_path.as_posix(),
-        "inputs": inputs,
-        "operation_context": {
-            "request_id": "test-request-id",
-            "user_agent": "test-user-agent",
-        },
-    }
+from .....utils import construct_flow_execution_request_json, load_content
 
 
 @pytest.mark.unittest
@@ -57,3 +36,9 @@ class TestExecutionApis:
         keywords_not_in_log = [f"Failed to execute flow, flow run id: {run_id}. Error:"]
         assert all(word in logs for word in keywords_in_log)
         assert all(word not in logs for word in keywords_not_in_log)
+
+    def test_cancel_execution(self, executor_client: TestClient):
+        request = {"run_id": "test-run-id"}
+        response = executor_client.post(url="/execution/cancel", json=request)
+        assert response.status_code == 200
+        assert response.json() == {"status": "canceled"}