+++ src/promptflow/tests/sdk_cli_test/e2etests/test_experiment.py
@@ -12,10 +12,10 @@ import pytest
 from mock import mock
 from ruamel.yaml import YAML
 
-from promptflow import PFClient
 from promptflow._sdk._constants import PF_TRACE_CONTEXT, ExperimentStatus, RunStatus, RunTypes
 from promptflow._sdk._errors import ExperimentValueError, RunOperationError
-from promptflow._sdk._load_functions import load_common
+from promptflow._sdk._load_functions import _load_experiment, load_common
+from promptflow._sdk._pf_client import PFClient
 from promptflow._sdk._submitter.experiment_orchestrator import ExperimentOrchestrator
 from promptflow._sdk.entities._experiment import CommandNode, Experiment, ExperimentTemplate, FlowNode
 
@@ -91,10 +91,10 @@ class TestExperiment:
         session = str(uuid.uuid4())
         if is_live():
             # Async start
-            exp = client._experiments.start(exp.name, session=session)
+            exp = client._experiments.start(exp, session=session)
             # Test the experiment in progress cannot be started.
             with pytest.raises(RunOperationError) as e:
-                client._experiments.start(exp.name)
+                client._experiments.start(exp)
             assert f"Experiment {exp.name} is {exp.status}" in str(e.value)
             assert exp.status in [ExperimentStatus.IN_PROGRESS, ExperimentStatus.QUEUING]
             exp = self.wait_for_experiment_terminated(client, exp)
@@ -123,10 +123,10 @@ class TestExperiment:
             assert len(line_runs) == 3
             line_run = line_runs[0]
             assert len(line_run.evaluations) == 1, "line run evaluation not exists!"
-            assert "eval_classification_accuracy" == line_run.evaluations[0].display_name
+            assert "eval_classification_accuracy" == list(line_run.evaluations.values())[0].display_name
 
         # Test experiment restart
-        exp = client._experiments.start(exp.name)
+        exp = client._experiments.start(exp)
         exp = self.wait_for_experiment_terminated(client, exp)
         for name, runs in exp.node_runs.items():
             assert all([run["status"] == RunStatus.COMPLETED] for run in runs)
@@ -142,7 +142,7 @@ class TestExperiment:
         exp = client._experiments.create_or_update(experiment)
         if is_live():
             # Async start
-            exp = client._experiments.start(exp.name)
+            exp = client._experiments.start(exp)
             exp = self.wait_for_experiment_terminated(client, exp)
         else:
             exp = client._experiments.get(exp.name)
@@ -163,11 +163,11 @@ class TestExperiment:
         experiment = Experiment.from_template(template)
         client = PFClient()
         exp = client._experiments.create_or_update(experiment)
-        exp = client._experiments.start(exp.name)
+        exp = client._experiments.start(exp)
         exp = self.wait_for_experiment_terminated(client, exp)
 
         # Test start experiment from nodes
-        exp = client._experiments.start(exp.name, from_nodes=["main"])
+        exp = client._experiments.start(exp, from_nodes=["main"])
         exp = self.wait_for_experiment_terminated(client, exp)
 
         assert exp.status == ExperimentStatus.TERMINATED
@@ -179,7 +179,7 @@ class TestExperiment:
         assert len(exp.node_runs["echo"]) == 2
 
         # Test run nodes in experiment
-        exp = client._experiments.start(exp.name, nodes=["main"])
+        exp = client._experiments.start(exp, nodes=["main"])
         exp = self.wait_for_experiment_terminated(client, exp)
 
         assert exp.status == ExperimentStatus.TERMINATED
@@ -197,10 +197,10 @@ class TestExperiment:
         experiment = Experiment.from_template(template)
         client = PFClient()
         exp = client._experiments.create_or_update(experiment)
-        exp = client._experiments.start(exp.name)
+        exp = client._experiments.start(exp)
         assert exp.status in [ExperimentStatus.IN_PROGRESS, ExperimentStatus.QUEUING]
         sleep(10)
-        client._experiments.stop(exp.name)
+        client._experiments.stop(exp)
         exp = client._experiments.get(exp.name)
         assert exp.status == ExperimentStatus.TERMINATED
 
@@ -255,7 +255,7 @@ class TestExperiment:
                 assert len(line_runs) == 1
                 line_run = line_runs[0]
                 assert len(line_run.evaluations) == 1, "line run evaluation not exists!"
-                assert "eval_classification_accuracy" == line_run.evaluations[0].display_name
+                assert "eval_classification_accuracy" == list(line_run.evaluations.values())[0].display_name
             # Test with default data and custom path
             expected_output_path = Path(tempfile.gettempdir()) / ".promptflow/my_custom"
             result = client.flows.test(target_flow_path, experiment=template_path, output_path=expected_output_path)
@@ -295,3 +295,14 @@ class TestExperiment:
                 "main2": {"output": "Hello world! text"},
                 "main3": {"output": "Hello world! Hello world! text"},
             }
+
+    @pytest.mark.usefixtures("use_secrets_config_file", "recording_injection", "setup_local_connection")
+    def test_experiment_with_script_run(self):
+        experiment_path = EXP_ROOT / "basic-script-template" / "basic-script.exp.yaml"
+        experiment = _load_experiment(experiment_path)
+        client = PFClient()
+        exp = client._experiments.start(experiment, stream=True, inputs={"count": 3})
+        assert exp.status == ExperimentStatus.TERMINATED
+        assert len(exp.node_runs) == 4
+        for key, val in exp.node_runs.items():
+            assert val[0]["status"] == RunStatus.COMPLETED, f"Node {key} run failed"