+++ src/promptflow/tests/test_configs/eager_flows/primitive_output/primitive_output_entry.py
@@ -2,4 +2,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
 
-__path__ = __import__("pkgutil").extend_path(__path__, __name__)  # type: ignore
+def my_flow(input_val: str = "gpt") -> str:
+    """Simple flow without yaml."""
+    # print(f"Hello world! {input_val}")
+    return f"Hello world! {input_val}"