+++ src/promptflow/tests/test_configs/eager_flows/simple_with_dict_output/dict_output_entry.py
@@ -2,7 +2,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
 
-def my_flow(input_val: str = "gpt"):
+def my_flow(input_val: str = "gpt") -> dict:
     """Simple flow without yaml."""
-    # print(f"Hello world! {input_val}")
-    return f"Hello world! {input_val}"
+    return {"output": f"Hello world! {input_val}"}